Start refactoring the poll code. Instead of leaving everything in rpcblockchain, a new directory 'contracts' is created, which should hold all the code using the gridcoin Smart Contract structure.
