+++ src/Makefile.qt.include
@@ -332,7 +332,7 @@ qt_libgridcoinqt_a_SOURCES = $(GRIDCOINRESEARCH_QT_CPP) $(GRIDCOINRESEARCH_QT_H)
   $(QT_QRC) $(QT_QRC_LOCALE) $(QT_TS) $(PROTOBUF_PROTO) $(RES_ICONS) $(RES_IMAGES) $(RES_MOVIES) $(RES_STYLESHEETS)
 
 if TARGET_WINDOWS
-qt_libgridcoinqt_a_SOURCES += $(GRIDCOIN_CORE_H) $(GRIDCOIN_CORE_CPP) grcrestarter.cpp neuralnet.cpp
+qt_libgridcoinqt_a_SOURCES += $(GRIDCOIN_CORE_H) $(GRIDCOIN_CORE_CPP) $(GRIDCOIN_CONTRACT_H) $(GRIDCOIN_CONTRACT_CPP) grcrestarter.cpp neuralnet.cpp
 qt_libgridcoinqt_a_CXXFLAGS += -DUNICODE
 endif
