+++ src/rpcblockchain.cpp
@@ -17,6 +17,8 @@
 #include "backup.h"
 #include "appcache.h"
 #include "tally.h"
+#include "contract/polls.h"
+#include "contract/contract.h"
 
 #include <boost/filesystem.hpp>
 #include <iostream>
@@ -34,8 +36,6 @@ extern double DoubleFromAmount(int64_t amount);
 std::string PubKeyToAddress(const CScript& scriptPubKey);
 CBlockIndex* GetHistoricalMagnitude(std::string cpid);
 extern std::string GetProvableVotingWeightXML();
-extern double ReturnVerifiedVotingBalance(std::string sXML, bool bCreatedAfterSecurityUpgrade);
-extern double ReturnVerifiedVotingMagnitude(std::string sXML, bool bCreatedAfterSecurityUpgrade);
 bool AskForOutstandingBlocks(uint256 hashStart);
 bool WriteKey(std::string sKey, std::string sValue);
 bool ForceReorganizeToHash(uint256 NewHash);
@@ -48,23 +48,16 @@ std::string ConvertBinToHex(std::string a);
 std::string ConvertHexToBin(std::string a);
 extern std::vector<unsigned char> readFileToVector(std::string filename);
 bool bNetAveragesLoaded_retired;
-extern bool SignBlockWithCPID(const std::string& sCPID, const std::string& sBlockHash, std::string& sSignature, std::string& sError, bool bAdvertising = false);
 std::string BurnCoinsWithNewContract(bool bAdd, std::string sType, std::string sPrimaryKey, std::string sValue, int64_t MinimumBalance, double dFees, std::string strPublicKey, std::string sBurnAddress);
-extern std::string GetBurnAddress();
 bool StrLessThanReferenceHash(std::string rh);
-extern std::string AddMessage(bool bAdd, std::string sType, std::string sKey, std::string sValue, std::string sSig, int64_t MinimumBalance, double dFees, std::string sPublicKey);
 extern std::string ExtractValue(std::string data, std::string delimiter, int pos);
 extern Array SuperblockReport(std::string cpid);
 MiningCPID GetBoincBlockByIndex(CBlockIndex* pblockindex);
 extern double GetSuperblockMagnitudeByCPID(std::string data, std::string cpid);
-extern bool VerifyCPIDSignature(std::string sCPID, std::string sBlockHash, std::string sSignature);
 std::string GetQuorumHash(const std::string& data);
 double GetOutstandingAmountOwed(StructCPID &mag, std::string cpid, int64_t locktime, double& total_owed, double block_magnitude);
 bool UpdateNeuralNetworkQuorumData();
-extern std::string AddContract(std::string sType, std::string sName, std::string sContract);
-StructCPID GetLifetimeCPID(const std::string& cpid, const std::string& sFrom);
 int64_t GetEarliestWalletTransaction();
-extern bool CheckMessageSignature(std::string sAction,std::string messagetype, std::string sMsg, std::string sSig, std::string opt_pubkey);
 bool LoadAdminMessages(bool bFullTableScan,std::string& out_errors);
 int64_t GetMaximumBoincSubsidy(int64_t nTime);
 double GRCMagnitudeUnit(int64_t locktime);
@@ -100,18 +93,11 @@ extern std::string ExecuteRPCCommand(std::string method, std::string arg1, std::
 extern std::string ExecuteRPCCommand(std::string method, std::string arg1, std::string arg2, std::string arg3, std::string arg4, std::string arg5);
 bool GetEarliestStakeTime(std::string grcaddress, std::string cpid);
 
+StructCPID GetLifetimeCPID(const std::string& cpid, const std::string& sFrom);
 
-extern Array GetJSONPollsReport(bool bDetail, std::string QueryByTitle, std::string& out_export, bool bIncludeExpired);
-
-
-extern Array GetJsonVoteDetailsReport(std::string pollname);
 extern std::string GetPollXMLElementByPollTitle(std::string pollname, std::string XMLElement1, std::string XMLElement2);
 
-
-extern double PollDuration(std::string pollname);
-
 extern Array GetJSONBeaconReport();
-
 
 void GatherNeuralHashes();
 void qtSyncWithDPORNodes(std::string data);
@@ -177,6 +163,11 @@ std::string FileManifest()
     }
     return sMyManifest;
 }
+
+
+
+
+
 
 std::vector<unsigned char> readFileToVector(std::string filename)
 {
@@ -572,52 +563,6 @@ void fileopen_and_copy(std::string src, std::string dest)
     fclose(outfile);
 }
 
-std::string SignMessage(std::string sMsg, std::string sPrivateKey)
-{
-     CKey key;
-     std::vector<unsigned char> vchMsg = vector<unsigned char>(sMsg.begin(), sMsg.end());
-     std::vector<unsigned char> vchPrivKey = ParseHex(sPrivateKey);
-     std::vector<unsigned char> vchSig;
-     key.SetPrivKey(CPrivKey(vchPrivKey.begin(), vchPrivKey.end())); // if key is not correct openssl may crash
-     if (!key.Sign(Hash(vchMsg.begin(), vchMsg.end()), vchSig))
-     {
-             return "Unable to sign message, check private key.";
-     }
-
-     const std::string sig(vchSig.begin(), vchSig.end());
-     std::string SignedMessage = EncodeBase64(sig);
-     return SignedMessage;
-}
-
-bool CheckMessageSignature(std::string sAction,std::string messagetype, std::string sMsg, std::string sSig, std::string strMessagePublicKey)
-{
-     std::string strMasterPubKey = "";
-     if (messagetype=="project" || messagetype=="projectmapping")
-     {
-        strMasterPubKey= msMasterProjectPublicKey;
-     }
-     else
-     {
-         strMasterPubKey = msMasterMessagePublicKey;
-     }
-
-     if (!strMessagePublicKey.empty()) strMasterPubKey = strMessagePublicKey;
-     if (sAction=="D" && messagetype=="beacon") strMasterPubKey = msMasterProjectPublicKey;
-	 if (sAction=="D" && messagetype=="poll")   strMasterPubKey = msMasterProjectPublicKey;
-	 if (sAction=="D" && messagetype=="vote")   strMasterPubKey = msMasterProjectPublicKey;
-     if (messagetype=="constblkreward")         strMasterPubKey = msMasterProjectPublicKey;
-
-     std::string db64 = DecodeBase64(sSig);
-     CKey key;
-     if (!key.SetPubKey(ParseHex(strMasterPubKey))) return false;
-     std::vector<unsigned char> vchMsg = vector<unsigned char>(sMsg.begin(), sMsg.end());
-     std::vector<unsigned char> vchSig = vector<unsigned char>(db64.begin(), db64.end());
-     if (!key.Verify(Hash(vchMsg.begin(), vchMsg.end()), vchSig)) return false;
-     return true;
-
-}
-
-
 std::string ExtractValue(std::string data, std::string delimiter, int pos)
 {
     std::vector<std::string> vKeys = split(data.c_str(),delimiter);
@@ -629,8 +574,6 @@ std::string ExtractValue(std::string data, std::string delimiter, int pos)
 
     return keyvalue;
 }
-
-
 
 double GetAverageInList(std::string superblock,double& out_count)
 {
@@ -707,7 +650,6 @@ double GetSuperblockMagnitudeByCPID(std::string data, std::string cpid)
         return -1;
 }
 
-
 void GetSuperblockProjectCount(std::string data, double& out_project_count, double& out_whitelist_count)
 {
 	   // This is reserved in case we ever want to resync prematurely when the last superblock contains < .75% of whitelisted projects (remember we allow superblocks with up to .50% of the whitelisted projects, in case some project sites are being ddossed)
@@ -716,7 +658,6 @@ void GetSuperblockProjectCount(std::string data, double& out_project_count, doub
        out_whitelist_count = GetCountOf("project");
 	   if (fDebug10) LogPrintf(" GSPC:CountOfProjInBlock %f vs WhitelistedCount %f  \n",(double)out_project_count,(double)out_whitelist_count);
 }
-
 
 double GetSuperblockAvgMag(std::string data,double& out_beacon_count,double& out_participant_count,double& out_average, bool bIgnoreBeacons,int nHeight)
 {
@@ -754,8 +695,6 @@ double GetSuperblockAvgMag(std::string data,double& out_beacon_count,double& out
     }
 
 }
-
-
 
 bool TallyMagnitudesInSuperblock()
 {
@@ -861,14 +800,6 @@ bool TallyMagnitudesInSuperblock()
                 return false;
     }
 }
-
-std::string AddContract(std::string sType, std::string sName, std::string sContract)
-{
-            std::string sPass = (sType=="project" || sType=="projectmapping" || sType=="smart_contract") ? GetArgument("masterprojectkey", msMasterMessagePrivateKey) : msMasterMessagePrivateKey;
-            std::string result = AddMessage(true,sType,sName,sContract,sPass,AmountFromValue(1),.00001,"");
-            return result;
-}
-
 
 bool CPIDAcidTest2(std::string bpk, std::string externalcpid)
 {
@@ -2910,72 +2841,6 @@ Value writedata(const Array& params, bool fHelp)
 
 // Network RPC commands
 
-Value addpoll(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() != 6)
-        throw runtime_error(
-                "addpoll <title> <days> <question> <answers> <sharetype> <url>\n"
-                "\n"
-                "<title> -----> The title for poll with no spaces. Use _ in between words\n"
-                "<days> ------> The number of days the poll will run\n"
-                "<question> --> The question with no spaces. Use _ in between words\n"
-                "<answers> ---> The answers available for voter to choose from. Use - in between words and ; to seperate answers\n"
-                "<sharetype> -> The share type of the poll; 1 = Magnitude 2 = Balance 3 = Magnitude + Balance 4 = CPID count 5 = Participant count\n"
-                "<url> -------> The corresponding url for the poll\n"
-                "\n"
-                "Add a poll to the network; Requires 100K GRC balance\n");
-
-    Object res;
-
-    std::string Title = params[0].get_str();
-    double days = Round(params[1].get_real(), 0);
-    std::string Question = params[2].get_str();
-    std::string Answers = params[3].get_str();
-    double sharetype = Round(params[4].get_real(), 0);
-    std::string sURL = params[5].get_str();
-
-    if (Title.empty() || Question.empty() || Answers.empty() || sURL.empty())
-    {
-        res.push_back(Pair("Error", "Must specify a poll title, question, answers, and URL\n"));
-
-        return res;
-    }
-
-    if (days < 7)
-        res.push_back(Pair("Error", "Minimum duration is 7 days; please specify a longer poll duration."));
-
-    else
-    {
-        double nBalance = GetTotalBalance();
-
-        if (nBalance < 100000)
-            res.push_back(Pair("Error", "You must have a balance > 100,000 GRC to create a poll.  Please post the desired poll on https://cryptocurrencytalk.com/forum/464-gridcoin-grc/ or https://github.com/Erkan-Yilmaz/Gridcoin-tasks/issues/45"));
-
-        else
-        {
-            if (days < 0 || days == 0)
-                res.push_back(Pair("Error", "You must specify a positive value for days for the expiration date."));
-
-            else
-            {
-                if (sharetype != 1 && sharetype != 2 && sharetype != 3 && sharetype != 4 && sharetype != 5)
-                    res.push_back(Pair("Error", "You must specify a value of 1, 2, 3, 4 or 5 for the sharetype."));
-
-                else
-                {
-                    std::string expiration = RoundToString(GetAdjustedTime() + (days*86400), 0);
-                    std::string contract = "<TITLE>" + Title + "</TITLE><DAYS>" + RoundToString(days, 0) + "</DAYS><QUESTION>" + Question + "</QUESTION><ANSWERS>" + Answers + "</ANSWERS><SHARETYPE>" + ToString(sharetype) + "</SHARETYPE><URL>" + sURL + "</URL><EXPIRATION>" + expiration + "</EXPIRATION>";
-                    std::string result = AddContract("poll", Title,contract);
-
-                    res.push_back(Pair("Success", "Your poll has been added: " + result));
-                }
-            }
-        }
-    }
-
-    return res;
-}
-
 Value askforoutstandingblocks(const Array& params, bool fHelp)
 {
     if (fHelp || params.size() != 0)
@@ -3096,109 +2961,6 @@ Value encryptphrase(const Array& params, bool fHelp)
     return res;
 }
 
-Value listallpolls(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() != 0)
-        throw runtime_error(
-                "listallpolls\n"
-                "\n"
-                "Lists all polls\n");
-
-    LOCK(cs_main);
-
-    std::string out1;
-    Array res = GetJSONPollsReport(false, "", out1, true);
-
-    return res;
-}
-
-Value listallpolldetails(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() != 0)
-        throw runtime_error(
-                "listallpolldetails\n"
-                "\n"
-                "Lists all polls with details\n");
-
-    LOCK(cs_main);
-
-    std::string out1;
-    Array res = GetJSONPollsReport(true, "", out1, true);
-
-    return res;
-}
-
-Value listpolldetails(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() != 0)
-        throw runtime_error(
-                "listpolldetails\n"
-                "\n"
-                "Lists poll details\n");
-
-    LOCK(cs_main);
-
-    std::string out1;
-    Array res = GetJSONPollsReport(true, "", out1, false);
-
-    return res;
-}
-
-Value listpollresults(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() > 2 || params.size() < 1)
-        throw runtime_error(
-                "listpollresults <pollname> [bool:showexpired]\n"
-                "\n"
-                "<pollname> ----> name of the poll\n"
-                "[showexpired] -> Optional; Default false\n"
-                "\n"
-                "Displays results for specified poll\n");
-
-    LOCK(cs_main);
-
-    Array res;
-    bool bIncExpired = false;
-
-    if (params.size() == 2)
-        bIncExpired = params[1].get_bool();
-
-    std::string Title1 = params[0].get_str();
-
-    if (!PollExists(Title1))
-    {
-        Object result;
-
-        result.push_back(Pair("Error", "Poll does not exist.  Please listpolls."));
-        res.push_back(result);
-    }
-    else
-    {
-        std::string Title = params[0].get_str();
-        std::string out1 = "";
-        Array myPolls = GetJSONPollsReport(true, Title, out1, bIncExpired);
-        res.push_back(myPolls);
-    }
-
-    return res;
-}
-
-Value listpolls(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() != 0)
-        throw runtime_error(
-                "listpolls\n"
-                "\n"
-                "Lists polls\n");
-
-    LOCK(cs_main);
-
-    std::string out1;
-    Array res = GetJSONPollsReport(false, "", out1, false);
-
-    return res;
-}
-
 Value memorypool(const Array& params, bool fHelp)
 {
     if (fHelp || params.size() != 0)
@@ -3283,151 +3045,6 @@ Value restorepoint(const Array& params, bool fHelp)
     return res;
 }
 #endif
-
-Value vote(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() != 2)
-        throw runtime_error(
-                "vote <title> <answers>\n"
-                "\n"
-                "<title -> Title of poll being voted on\n"
-                "<answers> -> Answers chosen for specified poll seperated by ;\n"
-                "\n"
-                "Vote on a specific poll with specified answers\n");
-
-    Object res;
-
-    std::string Title = params[0].get_str();
-    std::string Answer = params[1].get_str();
-
-    if (Title.empty() || Answer.empty())
-    {
-        res.push_back(Pair("Error", "Must specify a poll title and answers\n"));
-
-        return res;
-    }
-
-    LOCK2(cs_main, pwalletMain->cs_wallet);
-
-    //Verify the Existence of the poll, the acceptability of the answer, and the expiration of the poll: (EXIST, EXPIRED, ACCEPTABLE)
-    //If polltype == 1, use magnitude, if 2 use Balance, if 3 use hybrid:
-    //6-20-2015
-    double nBalance = GetTotalBalance();
-    uint256 hashRand = GetRandHash();
-    std::string email = GetArgument("email", "NA");
-    boost::to_lower(email);
-    GlobalCPUMiningCPID.email = email;
-    GlobalCPUMiningCPID.cpidv2 = ComputeCPIDv2(GlobalCPUMiningCPID.email, GlobalCPUMiningCPID.boincruntimepublickey, hashRand);
-    GlobalCPUMiningCPID.lastblockhash = GlobalCPUMiningCPID.cpidhash;
-
-    if (!PollExists(Title))
-        res.push_back(Pair("Error", "Poll does not exist."));
-
-    else
-    {
-        if (PollExpired(Title))
-            res.push_back(Pair("Error", "Sorry, Poll is already expired."));
-
-        else
-        {
-            if (!PollAcceptableAnswer(Title, Answer))
-            {
-                std::string acceptable_answers = PollAnswers(Title);
-                res.push_back(Pair("Error", "Sorry, Answer " + Answer + " is not one of the acceptable answers, allowable answers are: " + acceptable_answers + ".  If you are voting multiple choice, please use a semicolon delimited vote string such as : 'dog;cat'."));
-            }
-
-            else
-            {
-                std::string sParam = SerializeBoincBlock(GlobalCPUMiningCPID, pindexBest->nVersion);
-                std::string GRCAddress = DefaultWalletAddress();
-                StructCPID structMag = GetInitializedStructCPID2(GlobalCPUMiningCPID.cpid, mvMagnitudes);
-                double dmag = structMag.Magnitude;
-                double poll_duration = PollDuration(Title) * 86400;
-
-                // Prevent Double Voting
-                std::string cpid1 = GlobalCPUMiningCPID.cpid;
-                std::string GRCAddress1 = DefaultWalletAddress();
-                GetEarliestStakeTime(GRCAddress1, cpid1);
-                double cpid_age = GetAdjustedTime() - ReadCache("global", "nCPIDTime").timestamp;
-                double stake_age = GetAdjustedTime() - ReadCache("global", "nGRCTime").timestamp;
-
-                StructCPID structGRC = GetInitializedStructCPID2(GRCAddress, mvMagnitudes);
-
-                LogPrintf("CPIDAge %f, StakeAge %f, Poll Duration %f \r\n", cpid_age, stake_age, poll_duration);
-
-                double dShareType= RoundFromString(GetPollXMLElementByPollTitle(Title, "<SHARETYPE>", "</SHARETYPE>"), 0);
-
-                // Share Type 1 == "Magnitude"
-                // Share Type 2 == "Balance"
-                // Share Type 3 == "Both"
-                if (cpid_age < poll_duration) dmag = 0;
-
-                if (stake_age < poll_duration) nBalance = 0;
-
-                if ((dShareType == 1) && cpid_age < poll_duration)
-                    res.push_back(Pair("Error", "Sorry, When voting in a magnitude poll, your CPID must be older than the poll duration."));
-
-                else if (dShareType == 2 && stake_age < poll_duration)
-                    res.push_back(Pair("Error", "Sorry, When voting in a Balance poll, your stake age must be older than the poll duration."));
-
-                else if (dShareType == 3 && stake_age < poll_duration && cpid_age < poll_duration)
-                    res.push_back(Pair("Error", "Sorry, When voting in a Both Share Type poll, your stake age Or your CPID age must be older than the poll duration."));
-
-                else
-                {
-                    std::string voter = "<CPIDV2>"+GlobalCPUMiningCPID.cpidv2 + "</CPIDV2><CPID>"
-                            + GlobalCPUMiningCPID.cpid + "</CPID><GRCADDRESS>" + GRCAddress + "</GRCADDRESS><RND>"
-                            + hashRand.GetHex() + "</RND><BALANCE>" + RoundToString(nBalance,2)
-                            + "</BALANCE><MAGNITUDE>" + RoundToString(dmag,0) + "</MAGNITUDE>";
-                    // Add the provable balance and the provable magnitude - this goes into effect July 1 2017
-                    voter += GetProvableVotingWeightXML();
-                    std::string pk = Title + ";" + GRCAddress + ";" + GlobalCPUMiningCPID.cpid;
-                    std::string contract = "<TITLE>" + Title + "</TITLE><ANSWER>" + Answer + "</ANSWER>" + voter;
-                    std::string result = AddContract("vote",pk,contract);
-                    std::string narr = "Your CPID weight is " + RoundToString(dmag,0) + " and your Balance weight is " + RoundToString(nBalance,0) + ".";
-
-                    res.push_back(Pair("Success", narr + " " + "Your vote has been cast for topic " + Title + ": With an Answer of " + Answer + ": " + result.c_str()));
-                }
-            }
-        }
-    }
-
-    return res;
-}
-
-Value votedetails(const Array& params, bool fHelp)
-{
-    if (fHelp || params.size() != 1)
-        throw runtime_error(
-                "votedetails <pollname>]n"
-                "\n"
-                "<pollname> Specified poll name\n"
-                "\n"
-                "Displays vote details of a specified poll\n");
-
-    Array res;
-
-    std::string Title = params[0].get_str();
-
-    if (!PollExists(Title))
-    {
-        Object results;
-
-        results.push_back(Pair("Error", "Poll does not exist.  Please listpolls."));
-        res.push_back(results);
-    }
-
-    else
-    {
-        LOCK(cs_main);
-
-        Array myVotes = GetJsonVoteDetailsReport(Title);
-
-        res.push_back(myVotes);
-    }
-
-    return res;
-}
 
 Value execute(const Array& params, bool fHelp)
 {
@@ -3623,405 +3240,10 @@ double GetMagnitudeByCpidFromLastSuperblock(std::string sCPID)
         return 0;
 }
 
-bool VerifyCPIDSignature(std::string sCPID, std::string sBlockHash, std::string sSignature)
-{
-    std::string sBeaconPublicKey = GetBeaconPublicKey(sCPID, false);
-    std::string sConcatMessage = sCPID + sBlockHash;
-    bool bValid = CheckMessageSignature("R","cpid", sConcatMessage, sSignature, sBeaconPublicKey);
-    if(!bValid)
-        LogPrintf("VerifyCPIDSignature: invalid signature sSignature=%s, cached key=%s\n"
-        ,sSignature, sBeaconPublicKey);
-
-    return bValid;
-}
-
-bool SignBlockWithCPID(const std::string& sCPID, const std::string& sBlockHash, std::string& sSignature, std::string& sError, bool bAdvertising)
-{
-    // Check if there is a beacon for this user
-    // If not then return false as GetStoresBeaconPrivateKey grabs from the config
-    if (!HasActiveBeacon(sCPID) && !bAdvertising)
-    {
-        sError = "No active beacon";
-        return false;
-    }
-    // Returns the Signature of the CPID+BlockHash message.
-    std::string sPrivateKey = GetStoredBeaconPrivateKey(sCPID);
-    std::string sMessage = sCPID + sBlockHash;
-    sSignature = SignMessage(sMessage,sPrivateKey);
-    // If we failed to sign then return false
-    if (sSignature == "Unable to sign message, check private key.")
-    {
-        sError = sSignature;
-        sSignature = "";
-        return false;
-    }
-
-    return true;
-}
-
-std::string GetPollContractByTitle(std::string objecttype, std::string title)
-{
-    for(const auto& item : ReadCacheSection(objecttype))
-    {
-        const std::string& contract = item.second.value;
-        const std::string& PollTitle = ExtractXML(contract,"<TITLE>","</TITLE>");
-        if(boost::iequals(PollTitle, title))
-            return contract;
-    }
-
-    return std::string();
-}
-
-bool PollExists(std::string pollname)
-{
-    std::string contract = GetPollContractByTitle("poll",pollname);
-    return contract.length() > 10 ? true : false;
-}
-
-bool PollExpired(std::string pollname)
-{
-    std::string contract = GetPollContractByTitle("poll",pollname);
-    double expiration = RoundFromString(ExtractXML(contract,"<EXPIRATION>","</EXPIRATION>"),0);
-    return (expiration < (double)GetAdjustedTime()) ? true : false;
-}
-
-
-bool PollCreatedAfterSecurityUpgrade(std::string pollname)
-{
-	// If the expiration is after July 1 2017, use the new security features.
-	std::string contract = GetPollContractByTitle("poll",pollname);
-	double expiration = RoundFromString(ExtractXML(contract,"<EXPIRATION>","</EXPIRATION>"),0);
-	return (expiration > 1498867200) ? true : false;
-}
-
-
-double PollDuration(std::string pollname)
-{
-    std::string contract = GetPollContractByTitle("poll",pollname);
-    double days = RoundFromString(ExtractXML(contract,"<DAYS>","</DAYS>"),0);
-    return days;
-}
-
 double DoubleFromAmount(int64_t amount)
 {
     return (double)amount / (double)COIN;
 }
-
-
-double GetMoneySupplyFactor()
-{
-
-        StructCPID structcpid = mvNetwork["NETWORK"];
-        double TotalCPIDS = mvMagnitudes.size();
-        double AvgMagnitude = structcpid.NetworkAvgMagnitude;
-        double TotalNetworkMagnitude = TotalCPIDS*AvgMagnitude;
-        if (TotalNetworkMagnitude < 100) TotalNetworkMagnitude=100;
-        double MoneySupply = DoubleFromAmount(pindexBest->nMoneySupply);
-        double Factor = (MoneySupply/TotalNetworkMagnitude+.01);
-        return Factor;
-
-}
-
-double PollCalculateShares(std::string contract, double sharetype, double MoneySupplyFactor, unsigned int VoteAnswerCount)
-{
-    std::string address = ExtractXML(contract,"<GRCADDRESS>","</GRCADDRESS>");
-    std::string cpid = ExtractXML(contract,"<CPID>","</CPID>");
-   	double magnitude = ReturnVerifiedVotingMagnitude(contract,PollCreatedAfterSecurityUpgrade(contract));
-	double balance = ReturnVerifiedVotingBalance(contract,PollCreatedAfterSecurityUpgrade(contract));
-	if (VoteAnswerCount < 1) VoteAnswerCount=1;
-    if (sharetype==1) return magnitude/VoteAnswerCount;
-    if (sharetype==2) return balance/VoteAnswerCount;
-
-    if (sharetype==3)
-    {
-        // https://github.com/gridcoin/Gridcoin-Research/issues/87#issuecomment-253999878
-        // Researchers weight is Total Money Supply / 5.67 * Magnitude
-        double UserWeightedMagnitude = (MoneySupplyFactor/5.67) * magnitude;
-        return (UserWeightedMagnitude+balance) / VoteAnswerCount;
-    }
-    if (sharetype==4)
-    {
-        if (magnitude > 0) return 1;
-        return 0;
-    }
-    if (sharetype==5)
-    {
-        if (address.length() > 5) return 1;
-        return 0;
-    }
-    return 0;
-}
-
-
-double VotesCount(std::string pollname, std::string answer, double sharetype, double& out_participants)
-{
-    double total_shares = 0;
-    out_participants = 0;
-
-    double MoneySupplyFactor = GetMoneySupplyFactor();
-
-    for(const auto& item : ReadCacheSection("vote"))
-    {
-        const std::string& contract = item.second.value;
-        const std::string& Title = ExtractXML(contract,"<TITLE>","</TITLE>");
-        const std::string& VoterAnswer = ExtractXML(contract,"<ANSWER>","</ANSWER>");
-        const std::vector<std::string>& vVoterAnswers = split(VoterAnswer.c_str(),";");
-        for (const std::string& voterAnswers : vVoterAnswers)
-        {
-            if (boost::iequals(pollname, Title) && boost::iequals(answer, voterAnswers))
-            {
-                double shares = PollCalculateShares(contract, sharetype, MoneySupplyFactor, vVoterAnswers.size());
-                total_shares += shares;
-                out_participants += 1.0 / vVoterAnswers.size();
-            }
-        }
-    }
-
-    return total_shares;
-}
-
-
-
-std::string GetPollXMLElementByPollTitle(std::string pollname, std::string XMLElement1, std::string XMLElement2)
-{
-    std::string contract = GetPollContractByTitle("poll",pollname);
-    std::string sElement = ExtractXML(contract,XMLElement1,XMLElement2);
-    return sElement;
-}
-
-
-bool PollAcceptableAnswer(std::string pollname, std::string answer)
-{
-    std::string contract = GetPollContractByTitle("poll",pollname);
-    std::string answers = ExtractXML(contract,"<ANSWERS>","</ANSWERS>");
-    std::vector<std::string> vAnswers = split(answers.c_str(),";");
-
-    //Allow multiple choice voting:
-    std::vector<std::string> vUserAnswers = split(answer.c_str(),";");
-    for (unsigned int x = 0; x < vUserAnswers.size(); x++)
-    {
-        bool bFoundAnswer = false;
-        for (unsigned int i = 0; i < vAnswers.size(); i++)
-        {
-                boost::to_lower(vAnswers[i]); //Contains Poll acceptable answers
-                std::string sUserAnswer = vUserAnswers[x];
-                boost::to_lower(sUserAnswer);
-                if (sUserAnswer == vAnswers[i])
-                {
-                        bFoundAnswer=true;
-                        break;
-                }
-        }
-        if (!bFoundAnswer) return false;
-    }
-    return true;
-}
-
-std::string PollAnswers(std::string pollname)
-{
-    std::string contract = GetPollContractByTitle("poll",pollname);
-    std::string answers = ExtractXML(contract,"<ANSWERS>","</ANSWERS>");
-    return answers;
-
-}
-
-
-std::string GetShareType(double dShareType)
-{
-    if (dShareType == 1) return "Magnitude";
-    if (dShareType == 2) return "Balance";
-    if (dShareType == 3) return "Magnitude+Balance";
-    if (dShareType == 4) return "CPID Count";
-    if (dShareType == 5) return "Participants";
-    return "?";
-}
-
-
-
-
-
-std::string GetProvableVotingWeightXML()
-{
-    std::string sXML = "<PROVABLEMAGNITUDE>";
-    //Retrieve the historical magnitude
-    if (IsResearcher(msPrimaryCPID))
-    {
-        StructCPID st1 = GetLifetimeCPID(msPrimaryCPID,"ProvableMagnitude()");
-        CBlockIndex* pHistorical = GetHistoricalMagnitude(msPrimaryCPID);
-        if (pHistorical->nHeight > 1 && pHistorical->nMagnitude > 0)
-        {
-            std::string sBlockhash = pHistorical->GetBlockHash().GetHex();
-            std::string sError;
-            std::string sSignature;
-            bool bResult = SignBlockWithCPID(msPrimaryCPID, pHistorical->GetBlockHash().GetHex(), sSignature, sError);
-            // Just because below comment it'll keep in line with that
-            if (!bResult)
-                sSignature = sError;
-            // Find the Magnitude from the last staked block, within the last 6 months, and ensure researcher has a valid current beacon (if the beacon is expired, the signature contain an error message)
-            sXML += "<CPID>" + msPrimaryCPID + "</CPID><INNERMAGNITUDE>"
-                    + RoundToString(pHistorical->nMagnitude,2) + "</INNERMAGNITUDE>" +
-                    "<HEIGHT>" + ToString(pHistorical->nHeight)
-                    + "</HEIGHT><BLOCKHASH>" + sBlockhash + "</BLOCKHASH><SIGNATURE>" + sSignature + "</SIGNATURE>";
-        }
-    }
-    sXML += "</PROVABLEMAGNITUDE>";
-
-    vector<COutput> vecOutputs;
-    pwalletMain->AvailableCoins(vecOutputs, false, NULL, true);
-    std::string sRow = "";
-    double dTotal = 0;
-    double dBloatThreshhold = 100;
-    double dCurrentItemCount = 0;
-    double dItemBloatThreshhold = 50;
-    // Iterate unspent coins from transactions owned by me that total over 100GRC (this prevents XML bloat)
-    sXML += "<PROVABLEBALANCE>";
-    for (auto const& out : vecOutputs)
-    {
-        int64_t nValue = out.tx->vout[out.i].nValue;
-        const CScript& pk = out.tx->vout[out.i].scriptPubKey;
-        Object entry;
-        CTxDestination address;
-        if (ExtractDestination(out.tx->vout[out.i].scriptPubKey, address))
-        {
-            if (CoinToDouble(nValue) > dBloatThreshhold)
-            {
-                std::string sScriptPubKey1 = HexStr(pk.begin(), pk.end());
-                std::string strAddress=CBitcoinAddress(address).ToString();
-                CKeyID keyID;
-                const CBitcoinAddress& bcAddress = CBitcoinAddress(address);
-                if (bcAddress.GetKeyID(keyID))
-                {
-                    bool IsCompressed;
-                    CKey vchSecret;
-                    if (pwalletMain->GetKey(keyID, vchSecret))
-                    {
-                        // Here we use the secret key to sign the coins, then we abandon the key.
-                        CSecret csKey = vchSecret.GetSecret(IsCompressed);
-                        CKey keyInner;
-                        keyInner.SetSecret(csKey,IsCompressed);
-                        std::string private_key = CBitcoinSecret(csKey,IsCompressed).ToString();
-                        std::string public_key = HexStr(keyInner.GetPubKey().Raw());
-                        std::vector<unsigned char> vchSig;
-                        keyInner.Sign(out.tx->GetHash(), vchSig);
-                        // Sign the coins we own
-                        std::string sSig(vchSig.begin(), vchSig.end());
-                        // Increment the total balance weight voting ability
-                        dTotal += CoinToDouble(nValue);
-                        sRow = "<ROW><TXID>" + out.tx->GetHash().GetHex() + "</TXID>" +
-                                "<AMOUNT>" + RoundToString(CoinToDouble(nValue),2) + "</AMOUNT>" +
-                                "<POS>" + RoundToString((double)out.i,0) + "</POS>" +
-                                "<PUBKEY>" + public_key + "</PUBKEY>" +
-                                "<SCRIPTPUBKEY>" + sScriptPubKey1  + "</SCRIPTPUBKEY>" +
-                                "<SIG>" + EncodeBase64(sSig) + "</SIG>" +
-                                "<MESSAGE></MESSAGE></ROW>";
-                        sXML += sRow;
-                        dCurrentItemCount++;
-                        if (dCurrentItemCount >= dItemBloatThreshhold)
-                            break;
-                    }
-                }
-            }
-        }
-    }
-
-
-    sXML += "<TOTALVOTEDBALANCE>" + RoundToString(dTotal,2) + "</TOTALVOTEDBALANCE>";
-    sXML += "</PROVABLEBALANCE>";
-    return sXML;
-
-}
-
-
-double ReturnVerifiedVotingBalance(std::string sXML, bool bCreatedAfterSecurityUpgrade)
-{
-	std::string sPayload = ExtractXML(sXML,"<PROVABLEBALANCE>","</PROVABLEBALANCE>");
-	double dTotalVotedBalance = RoundFromString(ExtractXML(sPayload,"<TOTALVOTEDBALANCE>","</TOTALVOTEDBALANCE>"),2);
-	double dLegacyBalance = RoundFromString(ExtractXML(sXML,"<BALANCE>","</BALANCE>"),0);
-
-	if (fDebug10) LogPrintf(" \n Total Voted Balance %f, Legacy Balance %f \n",(float)dTotalVotedBalance,(float)dLegacyBalance);
-
-	if (!bCreatedAfterSecurityUpgrade) return dLegacyBalance;
-
-	double dCounted = 0;
-	std::vector<std::string> vXML= split(sPayload.c_str(),"<ROW>");
-	for (unsigned int x = 0; x < vXML.size(); x++)
-	{
-		// Prove the contents of the XML as a 3rd party
-		CTransaction tx2;
-		uint256 hashBlock = 0;
-		uint256 uTXID(ExtractXML(vXML[x],"<TXID>","</TXID>"));
-		std::string sAmt = ExtractXML(vXML[x],"<AMOUNT>","</AMOUNT>");
-		std::string sPos = ExtractXML(vXML[x],"<POS>","</POS>");
-		std::string sXmlSig = ExtractXML(vXML[x],"<SIG>","</SIG>");
-		std::string sXmlMsg = ExtractXML(vXML[x],"<MESSAGE>","</MESSAGE>");
-		std::string sScriptPubKeyXml = ExtractXML(vXML[x],"<SCRIPTPUBKEY>","</SCRIPTPUBKEY>");
-		int32_t iPos = RoundFromString(sPos,0);
-		std::string sPubKey = ExtractXML(vXML[x],"<PUBKEY>","</PUBKEY>");
-		if (!sPubKey.empty() && !sAmt.empty() && !sPos.empty() && uTXID > 0)
-		{
-			if (GetTransaction(uTXID, tx2, hashBlock))
-			{
-				if (iPos >= 0 && iPos < (int32_t) tx2.vout.size())
-				{
-					    int64_t nValue2 = tx2.vout[iPos].nValue;
-					    const CScript& pk2 = tx2.vout[iPos].scriptPubKey;
-					    CTxDestination address2;
-						std::string sVotedPubKey = HexStr(pk2.begin(), pk2.end());
-						std::string sVotedGRCAddress = CBitcoinAddress(address2).ToString();
-						std::string sCoinOwnerAddress = PubKeyToAddress(pk2);
-						double dAmount = CoinToDouble(nValue2);
-						if (ExtractDestination(tx2.vout[iPos].scriptPubKey, address2))
-						{
-							if (sScriptPubKeyXml == sVotedPubKey && RoundToString(dAmount,2) == sAmt)
-							{
-								Object entry;
-      					   		entry.push_back(Pair("Audited Amount",ValueFromAmount(nValue2)));
- 						    	std::string sDecXmlSig = DecodeBase64(sXmlSig);
-							    CKey keyVerify;
-							    if (keyVerify.SetPubKey(ParseHex(sPubKey)))
-								{
-									  	std::vector<unsigned char> vchMsg1 = vector<unsigned char>(sXmlMsg.begin(), sXmlMsg.end());
-										std::vector<unsigned char> vchSig1 = vector<unsigned char>(sDecXmlSig.begin(), sDecXmlSig.end());
-										bool bValid = keyVerify.Verify(uTXID,vchSig1);
-										// Unspent Balance is proven to be owned by the voters public key, count the vote
-										if (bValid) dCounted += dAmount;
-								}
-
-							}
-						}
-					}
-				}
-			}
-		}
-
-		return dCounted;
-}
-
-double ReturnVerifiedVotingMagnitude(std::string sXML, bool bCreatedAfterSecurityUpgrade)
-{
-	double dLegacyMagnitude  = RoundFromString(ExtractXML(sXML,"<MAGNITUDE>","</MAGNITUDE>"),2);
-	if (!bCreatedAfterSecurityUpgrade) return dLegacyMagnitude;
-
-	std::string sMagXML = ExtractXML(sXML,"<PROVABLEMAGNITUDE>","</PROVABLEMAGNITUDE>");
-	std::string sMagnitude = ExtractXML(sMagXML,"<INNERMAGNITUDE>","</INNERMAGNITUDE>");
-	std::string sXmlSigned = ExtractXML(sMagXML,"<SIGNATURE>","</SIGNATURE>");
-	std::string sXmlBlockHash = ExtractXML(sMagXML,"<BLOCKHASH>","</BLOCKHASH>");
-	std::string sXmlCPID = ExtractXML(sMagXML,"<CPID>","</CPID>");
-	if (!sXmlBlockHash.empty() && !sMagnitude.empty() && !sXmlSigned.empty())
-	{
-		CBlockIndex* pblockindexMagnitude = mapBlockIndex[uint256(sXmlBlockHash)];
-		if (pblockindexMagnitude)
-		{
-				bool fResult = VerifyCPIDSignature(sXmlCPID, sXmlBlockHash, sXmlSigned);
-				bool fAudited = (RoundFromString(RoundToString(pblockindexMagnitude->nMagnitude,2),0)==RoundFromString(sMagnitude,0) && fResult);
-				if (fAudited) return (double)pblockindexMagnitude->nMagnitude;
-		}
-	}
-	return 0;
-}
-
-
-
 
 Array GetJsonUnspentReport()
 {
@@ -4225,169 +3447,6 @@ Array GetJsonUnspentReport()
     return results;
 }
 
-
-
-
-
-Array GetJsonVoteDetailsReport(std::string pollname)
-{
-    double total_shares = 0;
-    double participants = 0;
-    double MoneySupplyFactor = GetMoneySupplyFactor();
-
-    Array results;
-    Object entry;
-    entry.push_back(Pair("Votes","Votes Report " + pollname));
-    entry.push_back(Pair("MoneySupplyFactor",RoundToString(MoneySupplyFactor,2)));
-
-    // Add header
-    entry.push_back(Pair("GRCAddress,CPID,Question,Answer,ShareType,URL", "Shares"));
-
-    boost::to_lower(pollname);
-    for(const auto& item : ReadCacheSection("vote"))
-    {
-        const std::string& contract = item.second.value;
-        const std::string& Title = ExtractXML(contract,"<TITLE>","</TITLE>");
-        if(boost::iequals(pollname, Title))
-        {
-            const std::string& OriginalContract = GetPollContractByTitle("poll",Title);
-            const std::string& Question = ExtractXML(OriginalContract,"<QUESTION>","</QUESTION>");
-            const std::string& GRCAddress = ExtractXML(contract,"<GRCADDRESS>","</GRCADDRESS>");
-            const std::string& CPID = ExtractXML(contract,"<CPID>","</CPID>");
-
-            double dShareType = RoundFromString(GetPollXMLElementByPollTitle(Title,"<SHARETYPE>","</SHARETYPE>"),0);
-            std::string sShareType= GetShareType(dShareType);
-            std::string sURL = ExtractXML(contract,"<URL>","</URL>");
-
-            std::string Balance = ExtractXML(contract,"<BALANCE>","</BALANCE>");
-
-            const std::string& VoterAnswer = boost::to_lower_copy(ExtractXML(contract,"<ANSWER>","</ANSWER>"));
-            const std::vector<std::string>& vVoterAnswers = split(VoterAnswer.c_str(),";");
-            for (const auto& answer : vVoterAnswers)
-            {
-                double shares = PollCalculateShares(contract, dShareType, MoneySupplyFactor, vVoterAnswers.size());
-                total_shares += shares;
-                participants += 1.0 / vVoterAnswers.size();
-                const std::string& voter = GRCAddress + "," + CPID + "," + Question + "," + answer + "," + sShareType + "," + sURL;
-                entry.push_back(Pair(voter,RoundToString(shares,0)));
-            }
-        }
-    }
-
-    entry.push_back(Pair("Total Participants",RoundToString(participants,2)));
-    results.push_back(entry);
-    return results;
-}
-
-
-Array GetJSONPollsReport(bool bDetail, std::string QueryByTitle, std::string& out_export, bool IncludeExpired)
-{
-    //Title,ExpirationDate, Question, Answers, ShareType(1=Magnitude,2=Balance,3=Both)
-    Array results;
-    Object entry;
-    entry.push_back(Pair("Polls","Polls Report " + QueryByTitle));
-    std::string rows;
-    std::string row;
-    double iPollNumber = 0;
-    double total_participants = 0;
-    double total_shares = 0;
-    boost::to_lower(QueryByTitle);
-    std::string sExport;
-    std::string sExportRow;
-    out_export.clear();
-
-    for(const auto& item : ReadCacheSection("poll"))
-    {
-        const std::string& title = boost::to_lower_copy(item.first);
-        const std::string& contract = item.second.value;
-        std::string Expiration = ExtractXML(contract,"<EXPIRATION>","</EXPIRATION>");
-        std::string Question = ExtractXML(contract,"<QUESTION>","</QUESTION>");
-        std::string Answers = ExtractXML(contract,"<ANSWERS>","</ANSWERS>");
-        std::string ShareType = ExtractXML(contract,"<SHARETYPE>","</SHARETYPE>");
-        std::string sURL = ExtractXML(contract,"<URL>","</URL>");
-        if (!PollExpired(title) || IncludeExpired)
-        {
-            if (QueryByTitle.empty() || QueryByTitle == title)
-            {
-
-                if( (title.length()>128) &&
-                    (Expiration.length()>64) &&
-                    (Question.length()>4096) &&
-                    (Answers.length()>8192) &&
-                    (ShareType.length()>64) &&
-                    (sURL.length()>256)  )
-                    continue;
-
-                const std::vector<std::string>& vAnswers = split(Answers.c_str(),";");
-
-                std::string::size_type longestanswer = 0;
-                for (const std::string& answer : vAnswers)
-                    longestanswer = std::max( longestanswer, answer.length() );
-
-                if( longestanswer>128 )
-                    continue;
-
-                iPollNumber++;
-                total_participants = 0;
-                total_shares=0;
-                std::string BestAnswer;
-                double highest_share = 0;
-                std::string ExpirationDate = TimestampToHRDate(RoundFromString(Expiration,0));
-                std::string sShareType = GetShareType(RoundFromString(ShareType,0));
-                std::string TitleNarr = "Poll #" + RoundToString((double)iPollNumber,0)
-                                        + " (" + ExpirationDate + " ) - " + sShareType;
-
-                entry.push_back(Pair(TitleNarr,title));
-                sExportRow = "<POLL><URL>" + sURL + "</URL><TITLE>" + title + "</TITLE><EXPIRATION>" + ExpirationDate + "</EXPIRATION><SHARETYPE>" + sShareType + "</SHARETYPE><QUESTION>" + Question + "</QUESTION><ANSWERS>"+Answers+"</ANSWERS>";
-
-                if (bDetail)
-                {
-                    entry.push_back(Pair("Question",Question));
-                    sExportRow += "<ARRAYANSWERS>";
-                    size_t i = 0;
-                    for (const std::string& answer : vAnswers)
-                    {
-                        double participants=0;
-                        double dShares = VotesCount(title, answer, RoundFromString(ShareType,0),participants);
-                        if (dShares > highest_share)
-                        {
-                            highest_share = dShares;
-                            BestAnswer = answer;
-                        }
-
-                        entry.push_back(Pair("#" + ToString(++i) + " [" + RoundToString(participants,3) + "]. " + answer,dShares));
-                        total_participants += participants;
-                        total_shares += dShares;
-                        sExportRow += "<RESERVED></RESERVED><ANSWERNAME>" + answer + "</ANSWERNAME><PARTICIPANTS>" + RoundToString(participants,0) + "</PARTICIPANTS><SHARES>" + RoundToString(dShares,0) + "</SHARES>";
-                    }
-                    sExportRow += "</ARRAYANSWERS>";
-
-                    //Totals:
-                    entry.push_back(Pair("Participants",total_participants));
-                    entry.push_back(Pair("Total Shares",total_shares));
-                    if (total_participants < 3) BestAnswer = "";
-
-                    entry.push_back(Pair("Best Answer",BestAnswer));
-                    sExportRow += "<TOTALPARTICIPANTS>" + RoundToString(total_participants,0)
-                                  + "</TOTALPARTICIPANTS><TOTALSHARES>" + RoundToString(total_shares,0)
-                                  + "</TOTALSHARES><BESTANSWER>" + BestAnswer + "</BESTANSWER>";
-
-                }
-                sExportRow += "</POLL>";
-                sExport += sExportRow;
-            }
-        }
-    }
-
-    results.push_back(entry);
-    out_export = sExport;
-    return results;
-}
-
-
-
-
-
 Array GetUpgradedBeaconReport()
 {
     Array results;
@@ -4415,10 +3474,6 @@ Array GetUpgradedBeaconReport()
     results.push_back(entry);
     return results;
 }
-
-
-
-
 
 Array GetJSONBeaconReport()
 {
@@ -4618,11 +3673,6 @@ Array GetJSONVersionReport()
 
 }
 
-std::string GetBurnAddress()
-{
-    return fTestNet ? "mk1e432zWKH1MW57ragKywuXaWAtHy1AHZ" : "S67nL4vELWwdDVzjgtEP4MxryarTZ9a8GB";
-}
-
 std::string BurnCoinsWithNewContract(bool bAdd, std::string sType, std::string sPrimaryKey, std::string sValue,
                      int64_t MinimumBalance, double dFees, std::string strPublicKey, std::string sBurnAddress)
 {
@@ -4662,32 +3712,6 @@ std::string SendReward(std::string sAddress, int64_t nAmount)
     wtx.hashBoinc = sMessageType + sMessageValue;
     string strError = pwalletMain->SendMoneyToDestinationWithMinimumBalance(address.Get(), nAmount, 1, wtx);
     if (!strError.empty()) return strError;
-    return wtx.GetHash().GetHex().c_str();
-}
-
-
-
-std::string AddMessage(bool bAdd, std::string sType, std::string sPrimaryKey, std::string sValue,
-                    std::string sMasterKey, int64_t MinimumBalance, double dFees, std::string strPublicKey)
-{
-    std::string sAddress = GetBurnAddress();
-    CBitcoinAddress address(sAddress);
-    if (!address.IsValid())       throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, "Invalid Gridcoin address");
-    int64_t nAmount = AmountFromValue(dFees);
-    // Wallet comments
-    CWalletTx wtx;
-    if (pwalletMain->IsLocked())  throw JSONRPCError(RPC_WALLET_UNLOCK_NEEDED, "Error: Please enter the wallet passphrase with walletpassphrase first.");
-    std::string sMessageType      = "<MT>" + sType  + "</MT>";  //Project or Smart Contract
-    std::string sMessageKey       = "<MK>" + sPrimaryKey   + "</MK>";
-    std::string sMessageValue     = "<MV>" + sValue + "</MV>";
-    std::string sMessagePublicKey = "<MPK>"+ strPublicKey + "</MPK>";
-    std::string sMessageAction    = bAdd ? "<MA>A</MA>" : "<MA>D</MA>"; //Add or Delete
-    //Sign Message
-    std::string sSig = SignMessage(sType+sPrimaryKey+sValue,sMasterKey);
-    std::string sMessageSignature = "<MS>" + sSig + "</MS>";
-    wtx.hashBoinc = sMessageType+sMessageKey+sMessageValue+sMessageAction+sMessagePublicKey+sMessageSignature;
-    string strError = pwalletMain->SendMoneyToDestinationWithMinimumBalance(address.Get(), nAmount, MinimumBalance, wtx);
-    if (!strError.empty())        throw JSONRPCError(RPC_WALLET_ERROR, strError);
     return wtx.GetHash().GetHex().c_str();
 }
